chore(visualization): move function for visualization to shared folder
