+++ src/timeMachine/components/visualization/view.tsx
@@ -3,13 +3,13 @@ import React, {FC, useCallback} from 'react'
 import {useDispatch, useSelector} from 'react-redux'
 
 // Components
-import ViewTypeDropdown from 'src/flows/pipes/Visualization/ViewTypeDropdown'
+import ViewTypeDropdown from 'src/timeMachine/components/visualization/ViewTypeDropdown'
 
 // Selectors
 import {getActiveTimeMachine} from 'src/timeMachine/selectors'
 
 // Types
-import { AppState } from 'src/types'
+import {AppState} from 'src/types'
 
 // Actions
 import {setType} from 'src/timeMachine/actions'
@@ -17,24 +17,21 @@ import {setType} from 'src/timeMachine/actions'
 export const Visualization: FC<{}> = () => {
   const dispatch = useDispatch()
 
-  const viewType = useSelector(
-    (state: AppState) => {
-      const {view} = getActiveTimeMachine(state)
+  const viewType = useSelector((state: AppState) => {
+    const {view} = getActiveTimeMachine(state)
 
-      return view.properties.type
-    }
+    return view.properties.type
+  })
+
+  const updateType = useCallback(
+    selectedType => {
+      dispatch(setType(selectedType))
+    },
+    [dispatch]
   )
 
-  const updateType =
-    useCallback((selectedType) => {
-      dispatch(setType(selectedType))
-    }, [dispatch])
-  
-  return ( 
-    <ViewTypeDropdown
-    viewType={viewType}
-    onUpdateType={updateType as any}
-  />
+  return (
+    <ViewTypeDropdown viewType={viewType} onUpdateType={updateType as any} />
   )
 }
